modifications for comparable
