+++ src/main/java/us/springett/parsers/cpe/ICpe.java
@@ -1,0 +1,261 @@
+/*
+ * This file is part of CPE Parser.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright (c) Steve Springett. All Rights Reserved.
+ */
+package us.springett.parsers.cpe;
+
+import us.springett.parsers.cpe.values.Part;
+import us.springett.parsers.cpe.exceptions.CpeEncodingException;
+
+/**
+ * Object representation of a Common Platform Enumeration (CPE).
+ *
+ * @author Jeremy Long
+ */
+public interface ICpe extends Comparable {
+
+    /**
+     * <p>
+     * Gets the part for the CPE entry.</p>
+     * <pre>cpe:2.3:<b>[part]</b>:[vendor]:[product]:[version]:[update]:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the part for the CPE entry
+     */
+    public Part getPart();
+
+    /**
+     * <p>
+     * Gets the vendor for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:<b>[vendor]</b>:[product]:[version]:[update]:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the vendor for the CPE entry
+     */
+    public String getVendor();
+
+    /**
+     * <p>
+     * Gets the product for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:<b>[product]</b>:[version]:[update]:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the product for the CPE entry
+     */
+    public String getProduct();
+
+    /**
+     * <p>
+     * Gets the version for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:<b>[version]</b>:[update]:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the version for the CPE entry
+     */
+    public String getVersion();
+
+    /**
+     * <p>
+     * Gets the update for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:<b>[update]</b>:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the update for the CPE entry
+     */
+    public String getUpdate();
+
+    /**
+     * <p>
+     * Gets the edition for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:<b>[edition]</b>:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the edition for the CPE entry
+     */
+    public String getEdition();
+
+    /**
+     * <p>
+     * Gets the language for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:[edition]:<b>[language]</b>:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the language for the CPE entry
+     */
+    public String getLanguage();
+
+    /**
+     * <p>
+     * Gets the software edition for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:[edition]:[language]:<b>[sw_edition]</b>:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the software edition for the CPE entry
+     */
+    public String getSwEdition();
+
+    /**
+     * <p>
+     * Gets the target software environment for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:[edition]:[language]:[sw_edition]:<b>[target_sw]</b>:[target_hw]:[other]</pre>
+     *
+     * @return the target software environment for the CPE entry.
+     */
+    public String getTargetSw();
+
+    /**
+     * <p>
+     * Gets the target hardware environment for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:[edition]:[language]:[sw_edition]:[target_sw]:<b>[target_hw]</b>:[other]</pre>
+     *
+     * @return the target hardware environment for the CPE entry
+     */
+    public String getTargetHw();
+
+    /**
+     * <p>
+     * Gets the other component for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:<b>[other]</b></pre>
+     *
+     * @return the other component for the CPE entry
+     */
+    public String getOther();
+
+    /**
+     * <p>
+     * Gets the well formed formatted vendor for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:<b>[vendor]</b>:[product]:[version]:[update]:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the vendor for the CPE entry
+     */
+    public String getWellFormedVendor();
+
+    /**
+     * <p>
+     * Gets the well formed formatted product for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:<b>[product]</b>:[version]:[update]:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the product for the CPE entry
+     */
+    public String getWellFormedProduct();
+
+    /**
+     * <p>
+     * Gets the well formed formatted version for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:<b>[version]</b>:[update]:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the version for the CPE entry
+     */
+    public String getWellFormedVersion();
+
+    /**
+     * <p>
+     * Gets the well formed formatted update for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:<b>[update]</b>:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the update for the CPE entry
+     */
+    public String getWellFormedUpdate();
+
+    /**
+     * <p>
+     * Gets the well formed formatted edition for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:<b>[edition]</b>:[language]:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the edition for the CPE entry
+     */
+    public String getWellFormedEdition();
+
+    /**
+     * <p>
+     * Gets the well formed formatted language for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:[edition]:<b>[language]</b>:[sw_edition]:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the language for the CPE entry
+     */
+    public String getWellFormedLanguage();
+
+    /**
+     * <p>
+     * Gets the well formed formatted software edition for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:[edition]:[language]:<b>[sw_edition]</b>:[target_sw]:[target_hw]:[other]</pre>
+     *
+     * @return the software edition for the CPE entry
+     */
+    public String getWellFormedSwEdition();
+
+    /**
+     * <p>
+     * Gets the well formed formatted target software environment for the CPE
+     * entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:[edition]:[language]:[sw_edition]:<b>[target_sw]</b>:[target_hw]:[other]</pre>
+     *
+     * @return the target software environment for the CPE entry.
+     */
+    public String getWellFormedTargetSw();
+
+    /**
+     * <p>
+     * Gets the well formed formatted target hardware environment for the CPE
+     * entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:[edition]:[language]:[sw_edition]:[target_sw]:<b>[target_hw]</b>:[other]</pre>
+     *
+     * @return the target hardware environment for the CPE entry
+     */
+    public String getWellFormedTargetHw();
+
+    /**
+     * <p>
+     * Gets the well formed formatted other component for the CPE entry.</p>
+     * <pre>cpe:2.3:[part]:[vendor]:[product]:[version]:[update]:[edition]:[language]:[sw_edition]:[target_sw]:[target_hw]:<b>[other]</b></pre>
+     *
+     * @return the other component for the CPE entry
+     */
+    public String getWellFormedOther();
+
+    /**
+     * Converts the CPE into the CPE 2.2 URI format.
+     *
+     * @return the CPE 2.2 URI format of the CPE
+     * @throws CpeEncodingException thrown if the CPE is not well formed
+     */
+    public String toCpe22Uri() throws CpeEncodingException;
+
+    /**
+     * Converts the CPE into the CPE 2.3 Formatted String.
+     *
+     * @return the CPE 2.3 Formatted String
+     */
+    public String toCpe23FS();
+
+    /**
+     * Determines if the CPE matches the given target CPE. This does not follow
+     * the CPE 2.3 Specification exactly as there are cases where undefined
+     * comparisons will result in either true or false. For instance, 'ANY' will
+     * match 'm+wild cards' and NA will return false when the target has 'm+wild
+     * cards'.
+     *
+     * @param target the target CPE to evaluate
+     * @return <code>true</code> if the CPE matches the target; otherwise
+     * <code>false</code>
+     */
+    public boolean matches(ICpe target);
+
+    /**
+     * Determines if the target CPE matches the CPE. This does not follow the
+     * CPE 2.3 Specification exactly as there are cases where undefined
+     * comparisons will result in either true or false. For instance, 'ANY' will
+     * match 'm+wild cards' and NA will return false when the target has 'm+wild
+     * cards'.
+     *
+     * @param target the CPE to evaluate
+     * @return <code>true</code> if the target CPE matches CPE; otherwise
+     * <code>false</code>
+     */
+    public boolean matchedBy(ICpe target);
+}
